# client/public/openapi/browsing.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"tags":[{"name":"Browsing Service","description":"INTERNAL API for browsing parliamentary protocols, speeches, and statistics. Its main purpose is to serve the client."}],"paths":{"/statistics":{"get":{"tags":["Browsing Service"],"summary":"Get system statistics","description":"Returns overall statistics about the parliamentary data system","operationId":"getStatistics","responses":{"200":{"description":"Statistics retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatisticsDto"}}}}}}},"/speeches":{"get":{"tags":["Browsing Service"],"summary":"Search and filter speeches","description":"Returns paginated speeches with optional filtering by parties, speakers, protocols, and text search with similarity matching","operationId":"getAllSpeeches","parameters":[{"name":"page","in":"query","description":"Page number (0-based)","required":false,"schema":{"type":"integer","format":"int32","default":0},"example":0},{"name":"size","in":"query","description":"Number of items per page","required":false,"schema":{"type":"integer","format":"int32","default":10},"example":10},{"name":"parties","in":"query","description":"Filter by party names","required":false,"schema":{"type":"array","items":{"type":"string"}},"example":["CDU","SPD"]},{"name":"speakerIds","in":"query","description":"Filter by speaker IDs","required":false,"schema":{"type":"array","items":{"type":"integer","format":"int32"}},"example":[1,2,3]},{"name":"plenaryProtocolId","in":"query","description":"Filter by plenary protocol ID","required":false,"schema":{"type":"integer","format":"int32"},"example":123},{"name":"searchText","in":"query","description":"Text to search for in speeches","required":false,"schema":{"type":"string"},"example":"climate change"},{"name":"searchSimilarityThreshold","in":"query","description":"Similarity threshold for text search (0.0-1.0)","required":false,"schema":{"type":"number","format":"float","default":0.5},"example":0.5}],"responses":{"200":{"description":"Speeches retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page"}}}}}}},"/speakers":{"get":{"tags":["Browsing Service"],"summary":"Get speaker statistics","description":"Returns paginated statistics for all speakers in the parliamentary system","operationId":"getAllSpeakerStatistics","parameters":[{"name":"page","in":"query","description":"Page number (0-based)","required":false,"schema":{"type":"integer","format":"int32","default":0},"example":0},{"name":"size","in":"query","description":"Number of items per page","required":false,"schema":{"type":"integer","format":"int32","default":10},"example":10}],"responses":{"200":{"description":"Speaker statistics retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page"}}}}}}},"/speaker/{id}/name":{"get":{"tags":["Browsing Service"],"summary":"Get speaker name by ID","description":"Returns the name of a specific speaker","operationId":"getSpeakerName","parameters":[{"name":"id","in":"path","description":"Speaker ID","required":true,"schema":{"type":"integer","format":"int32"},"example":456}],"responses":{"404":{"description":"Speaker not found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Speaker name retrieved successfully","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/plenary-protocols":{"get":{"tags":["Browsing Service"],"summary":"Get all plenary protocols","description":"Returns a paginated list of all plenary protocols with sorting options","operationId":"getAllPlenaryProtocols","parameters":[{"name":"page","in":"query","description":"Page number (0-based)","required":false,"schema":{"type":"integer","format":"int32","default":0},"example":0},{"name":"size","in":"query","description":"Number of items per page","required":false,"schema":{"type":"integer","format":"int32","default":10},"example":10},{"name":"sortBy","in":"query","description":"Field to sort by","required":false,"schema":{"type":"string","default":"date"},"example":"date"},{"name":"sortDir","in":"query","description":"Sort direction (asc or desc)","required":false,"schema":{"type":"string","default":"desc"},"example":"desc"}],"responses":{"200":{"description":"Plenary protocols retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page"}}}}}}},"/plenary-protocols/{id}/name":{"get":{"tags":["Browsing Service"],"summary":"Get plenary protocol name by ID","description":"Returns the name/title of a specific plenary protocol","operationId":"getPlenaryProtocolName","parameters":[{"name":"id","in":"path","description":"Plenary protocol ID","required":true,"schema":{"type":"integer","format":"int32"},"example":123}],"responses":{"404":{"description":"Protocol not found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Protocol name retrieved successfully","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/parties":{"get":{"tags":["Browsing Service"],"summary":"Get party statistics","description":"Returns statistics for all political parties","operationId":"getAllPartyStatistics","responses":{"200":{"description":"Party statistics retrieved successfully","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/health":{"get":{"tags":["Browsing Service"],"summary":"Health check endpoint","description":"Returns the health status of the service","operationId":"health","responses":{"200":{"description":"Service is healthy","content":{"application/json":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"StatisticsDto":{"type":"object","properties":{"plenaryCount":{"type":"integer","format":"int32"},"speakerCount":{"type":"integer","format":"int32"},"wordCount":{"type":"integer","format":"int32"},"partyCount":{"type":"integer","format":"int32"}}},"Page":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"type":"object"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}}}}}